Refer to Crossplane as a 'CNCF Project' - no 'Sandbox'
I can't find much about this online, but I recall that until our recent
graduation we were required to refer to Crossplane as a Sanbox project. I
believe the correct terminology for an incubating project is just "a CNCF
project" and have updated the site accordingly.

See https://www.cncf.io/services-for-projects/#marketing

Signed-off-by: Nic Cope <negz@rk0n.org>