Change wording for Github link to `Contribute on GitHub`
This could be confusing for first time visitors, as there is another path for 'Getting started', which is `Documentation > Getting Started`, which links [here](https://crossplane.io/docs/v1.10/getting-started.html). The GitHub repo itself as resources to contribute, but not on 'how to get started with Crossplane'.

I don't have data for this confusion happening often, but I know of one instance of a user that ended up on the wrong page.

So the new wording should separate the link to Github and the link to Documentation a bit more and lead potential contributors to Github and users to the Documentation page.

Signed-off-by: jonny.langefeld <jonny.langefeld@getcruise.com>