Clarify that this repo is only the docs
I was making a fresh clone of this repo when I saw that it was forked
from crossplane.github.io and stated in the README that it was both the
website and docs. This surprised me, since crossplane/docs seemed like a
misleading place to find the website.

I then learned that the website is in fact a separate repo, at
https://github.com/crossplane/website

Signed-off-by: Nic Cope <nicc@rk0n.org>