DO NOT MERGE: Operations braindump
This is an (AI assisted) braindump of all the bits of the Operations
implementation I think might be useful. I'm going on vacation and can't
write the docs myself, so I wanted to provide this for the folks who
are.

I've tried to err toward being overly verbose. We probably don't need to
document all this - e.g. a lot of it duplicates the generated API docs.

Signed-off-by: Nic Cope <nicc@rk0n.org>