Update header navigation community link to commercial
This was changed awhile ago on the main www.crossplane.io site to
more accurately reflect its intent, as permitted by the CNCF website
guidelines. This docs site header navigation should match that of
the main www site, and we missed making this symetrical update when
the main site was updated.

Signed-off-by: Jared Watts <jbw976@gmail.com>